# PERF: IndexEngine.get_indexer_non_unique