@mroeschke 2023-11-04
How would performance be if scaling followed 10000 + 2^n where 2^n was about 10000 on the first iteration?