@mroeschke 2023-11-04
If the resizing is the expensive operation here, I think this would be generically more performant for larger cases too